Update cli_i_value_t structure to change num_val type from size_t to long for better compatibility

+++ value/value.h
@@ -64,9 +64,9 @@ typedef struct
  */
 typedef struct
 {
-    const char *value;           /**< Pointer to the value of the command (might be NULL) */
+    const char *value;     /**< Pointer to the value of the command (might be NULL) */
     vector_t *vec_value;   /**< Vector of values for array flags/command (might be NULL) */
-    size_t num_val;        /**< Value for integer flags/commands (might be 0) */
+    long num_val;          /**< Value for integer flags/commands (might be 0) */
     float float_val;       /**< Value for float flags/commands (might be 0.0) */
 } cli_i_value_t;
